% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregatoR.R
\name{aggregatoR}
\alias{aggregatoR}
\title{Aggregation and summary statistics of time series}
\usage{
aggregatoR(
  x,
  time_col = "Date",
  time_format = "\%Y-\%m-\%d",
  stat_col = "Time_Series",
  statistic = mean,
  Summary = TRUE,
  season = FALSE,
  doy = c(74:180),
  export = FALSE,
  filename = NULL,
  ...
)
}
\arguments{
\item{x}{data.frame containing the time series to analyze; must inlcude a column with the timestamp}

\item{time_col}{Name of the column in which the information about time is stored}

\item{time_format}{specifies in which format the time is stored. See ?lubridate::parse_date_time() for further information}

\item{stat_col}{Name of the column in which the data to be aggregated or summarized is stored}

\item{statistic}{Statistic to use for aggregation, typically mean, median, min or max. Default is to use na.rm=TRUE when computing the statistics}

\item{season}{Boolean; if FALSE (the default), simple yearly aggregation is performed; if TRUE, aggregation is performed for the
period of the year specified by the range of days of the year given by doy}

\item{doy}{period defined by days of the year. Boundaries are included in the computation.}

\item{export}{boolean, if TRUE, the function returns a csv file containing the output.}

\item{filename}{filename (including path) of the output file if export=TRUE.}

\item{...}{Additional parameter for the statistic, e.g. a vector of probailities when computing quantiles as statistics}

\item{summary}{Boolean; if TRUE, summary statistics are computed during the aggregation}
}
\value{
A data.frame containing the aggregated or summarized time series
}
\description{
Helper function to bulk process of time series files; can be used to aggregate the series
specific season and to compute summary statistics.
}
\references{
Betz, F; Fischer, P. (in prep.): Assessing Spatial-Temporal Dynamics of Groundwater in a Restored Floodplain at the German Danube using Machine Learning. In Preparation.
}
\author{
Florian Betz
}
