% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation.R
\name{Animation}
\alias{Animation}
\title{Create Animated GIF from the Raster Time Series}
\usage{
Animation(
  raster_dir = NULL,
  raster_type = "tif",
  color_palette = "BluYl",
  range = c(370, 380),
  magick_path = "C:/Program Files/ImageMagick-7.0.10-Q16-HDRI",
  out_file = "animation.gif",
  ...
)
}
\arguments{
\item{raster_dir}{Directory, where the rasters for the animation are stored. Filenames are assumed to include the
timestamp (see details)}

\item{raster_type}{File type of the input rasters stored in raster_dir. Default format is tif}

\item{color_palette}{Color palette to use for plotting the single raster files. Any palette from the colorspace-package
is supported.}

\item{magick_path}{Parent directory of ImageMagick. Can be downloaded from https://imagemagick.org/}

\item{out_file}{Path to the resulting GIF file}

\item{...}{Further parameters for ani.options and saveGIF. See help("ani.options") and help("saveGIF") for further information.}
}
\value{
A GIF file with the animation of the raster time series as specified with 'out_file'
}
\description{
This function takes the entire raster time series and makes an animated GIF.
}
\details{
The functions assumes to take the output from the predict_rts() function from the GroundwatermodelR package
where file names include the timestamp: gwl_YYYYmmdd (gwd refers to groundwater level). For instance gwl_20201025.tif refers
to the interpolation raster on 2020-10-25. An option to include custom filename formats is not included yet.\cr
The function assumes 'magick.exe' to be present. See help("ani.options") for further details.
}
\references{
Betz, F; Fischer, P. (in prep.): Assessing Spatial-Temporal Dynamics of Groundwater in a Restored Floodplain at the German Danube using Machine Learning. In Preparation.
}
\author{
Florian Betz
}
