% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractR.R
\name{extractR}
\alias{extractR}
\title{Handles the resulting raster time series and extracts time series for specific locations}
\usage{
extractR(
  raster_dir = NULL,
  loc = NULL,
  out_dir = NULL,
  out_type = "df",
  raster_type = "tif",
  name_col = Name,
  dec = ","
)
}
\arguments{
\item{raster_dir}{Directory, where the rasters are stored. Filenames are assumed to include the
timestamp (see details)}

\item{loc}{path to any sf-compatible file containing either points or polyons for which to extract and summarize raster time series}

\item{out_dir}{The directory where to write the tables resulting from each location}

\item{out_type}{Create data.frame as output or write it to files. If "df", the function returns a data.frame in long format,
if out_type="export", the function exports one single csv-file per location to the directory specified in out_dir.}

\item{raster_type}{File type of the input rasters stored in raster_dir. Default format is tif}

\item{name_col}{Name of the attribute column in loc where the name or id of the location is stored. This can refer to a full
name or be simply an id. During the extraction, it is used to link the raster data clearly to the location.}

\item{dec}{the string to use for decimal points when writing the output to file. Default is "," as expected by MS Excel.}
}
\value{
A data.frame or csv files written to out_dir
}
\description{
This function is designed to work with the resulting raster time series and extract single time series as
table. It implements the idea of 'virtual gauges' described in Betz and Fischer (2023).
}
\details{
The functions assumes to take the output from the predictoR() function from the GroundwatermodelR package
where file names include the timestamp: gwl_YYYYmmdd (gwd refers to groundwater level). For instance gwl_20201025.tif refers
to the interpolation raster on 2020-10-25. An option to include custom filename formats is not included yet.\cr
The function can take points or polygons as input. In case of points, simply the values at the point location are returned.
In case of polygons, the function returns basic statistics for each time step of the values within the polygon. The
GroundwatermodelR package offers the function plotfindR to convert polygons to points.\cr
The function has two output options. Using out_type="df", it returns a data.frame containing all values in long format. If
out_type="export", the functions returns one .csv file for each location and stores it in the directory specified with out_dir.
}
\references{
Betz, F; Fischer, P. (in prep.): Assessing Spatial-Temporal Dynamics of Groundwater in a Restored Floodplain at the German Danube using Machine Learning. In Preparation.
}
\author{
Florian Betz
}
