% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroundwatermodelR-Package.R
\docType{package}
\name{GroundwatermodelR}
\alias{GroundwatermodelR}
\title{GroundwatermodelR}
\description{
Collection of functions for applying the CAST package for interpolation of groundwater level data. Includes
routines for data preprocessing and handling of the resulting raster time series.
}
\details{
Spatial-Temporal Interpolation of Groundwater Levels using the CAST Package
}
\references{
Betz, F; Fischer, P. (in prep.): Assessing Spatial-Temporal Dynamics of Groundwater in a Restored Floodplain at the German Danube using Machine Learning. In Preparation.
}
\author{
Florian Betz
}
