% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfindR.R
\name{plotfindR}
\alias{plotfindR}
\title{Finding suitable points within monitoring plots to use for groundwater distance computation}
\usage{
plotfindR(dem, loc, mode = "centroid")
}
\arguments{
\item{dem}{Path to a digital elevation model stored in any terra::rast() compatible raster format; has to be a single band raster}

\item{loc}{Path to the polyons representing the observation areas. Can be any sf::st_read() compatible format}

\item{mode}{Defines the location within the plot area to be returned and can be one of the following:
"lowest", "heighest", "centroid"; Default is "Centroid"}
}
\value{
Simple feature object containing the lowest/heighest points or centroids of each polygon of loc
}
\description{
This function uses a digital elevation model along with polygons representing e.g. vegetation plots to obtain
the plot centroid, lowest or heighest point of the plot. The output can then be used to speed up the computation in the
ExtractR function.
}
