% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictoR.R
\name{predictoR}
\alias{predictoR}
\title{Make predictions for new timesteps using existing model}
\usage{
predictoR(
  model = NULL,
  spat_predictors = NULL,
  temp_predictors = NULL,
  date_column = Date,
  date_format = "\%Y-\%m-\%d",
  out_basename = "gwl",
  out_format = "tif",
  out_dir = NULL,
  ncores = detectCores(),
  ...
)
}
\arguments{
\item{model}{Path to the existing model stored as .rds file}

\item{spat_predictors}{Raster file containing the spatial predictors of the model; has to be the same as used for model training}

\item{temp_predictors}{Data.frame containing the time series input for prediction where each row refers to one timestep}

\item{date_column}{Name of the column where the time information is stored in the table of temporal predictors; default is Date}

\item{date_format}{How the time information is formatted in date_column; please refer to as.Date() for}

\item{out_basename}{Basename of the output filenames. Default is "gwl" as this basename is used in the other functions of GroundwatermodelR}

\item{out_dir}{Directory where to store the raster files created during prediction}

\item{ncores}{number of cores to use during multicore prediction; default is one (i.e. no parallel processing), the number of
available on your machine can be estimated using parellel::detectCores()}

\item{...}{Additional parameter of terra::predict}
}
\value{
One raster file for each timestep in the temporal predictor stored in out_dir. The filename includes a timestamp.
}
\description{
This function uses an existing interpolation model
(e.g. trained using the TraineR function) and makes predictions for new timesteps
}
\references{
Betz, F; Fischer, P. (in prep.): Assessing Spatial-Temporal Dynamics of Groundwater in a Restored Floodplain at the German Danube using Machine Learning. In Preparation.
}
\author{
Florian Betz
}
